chrome/content/mochitest/tests/ui/test_icon_status.html

<!DOCTYPE HTML>
<html>
<head>
  <title>ABP icon status tests</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>        

  <script type="application/x-javascript;version=1.7" src="common.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none">

  </div>

  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    let abp = Components.classes["@mozilla.org/adblockplus;1"].createInstance().wrappedJSObject;

    let tests = [
      createToolbarTest(testEnabled),
      createToolbarTest(testDisabled),
      createToolbarTest(testWhitelisted),
      createStatusbarTest(testEnabled),
      createStatusbarTest(testDisabled),
      createStatusbarTest(testWhitelisted)
    ];
    let currentTest = -1;

    function testEnabled(element)
    {
      if (!abp.prefs.enabled)
      {
        abp.prefs.enabled = true;
        abp.prefs.save();
      }
      setTimeout(function()
      {
        is(element.getAttribute("abpstate"), "active", "Checking icon state when Adblock Plus is enabled");
        runNextTest();
      }, 0);
    }

    function testDisabled(element)
    {
      if (abp.prefs.enabled)
      {
        abp.prefs.enabled = false;
        abp.prefs.save();
      }
      setTimeout(function()
      {
        is(element.getAttribute("abpstate"), "disabled", "Checking icon state when Adblock Plus is disabled");
        runNextTest();
      }, 0);
    }

    function testWhitelisted(element)
    {
      abp.filterStorage.addFilter(abp.Filter.fromText("@@|chrome://mochikit/$document"));
      if (!abp.prefs.enabled)
      {
        abp.prefs.enabled = true;
        abp.prefs.save();
      }
      setTimeout(function()
      {
        is(element.getAttribute("abpstate"), "whitelisted", "Checking icon state with site whitelisting");
        abp.filterStorage.removeFilter(abp.Filter.fromText("@@|chrome://mochikit/$document"));
        runNextTest();
      }, 0);
    }

    function createToolbarTest(testFunction)
    {
      return function()
      {
        let icon = getBrowserWindow().document.getElementById("abp-toolbarbutton");
        if (icon)
          testFunction(icon, function() { abp.prefs.showintoolbar = oldShowInToolbar; });
        else
          ok(false, "Toolbar icon not found in browser");
      };
    }

    function createStatusbarTest(testFunction)
    {
      return function()
      {
        let icon = getBrowserWindow().document.getElementById("abp-status");
        if (icon)
          testFunction(icon);
        else
          ok(false, "Status bar icon not found in browser");
      };
    }

    function runNextTest()
    {
      currentTest++;
      if (currentTest >= tests.length)
        SimpleTest.finish();
      else
        tests[currentTest]();
    }

    function start()
    {
      window.addEventListener("unload", stop, false);

      abp.prefs.enabled = true;

      // Make sure status bar and toolbar icons are shown
      abp.prefs.showintoolbar = false;
      abp.prefs.showinstatusbar = false;
      abp.prefs.save();
      setTimeout(function()
      {
        abp.prefs.showintoolbar = true;
        abp.prefs.showinstatusbar = true;
        abp.prefs.save();

        setTimeout(runNextTest, 0);
      }, 0);
    }

    function stop()
    {
      abp.prefs.enabled = oldEnabled;
      abp.prefs.showintoolbar = oldShowInToolbar;
      abp.prefs.showinstatusbar = oldShowInStatusbar;
      abp.prefs.save();
    }

    let oldEnabled = abp.prefs.enabled;
    let oldShowInToolbar = abp.prefs.showintoolbar;
    let oldShowInStatusbar = abp.prefs.showinstatusbar;

    SimpleTest.waitForExplicitFinish();
    addLoadEvent(start);
  </script>
  </pre>
</body>
</html>
